Refactor Delivery::getPackages().
